chore: break up artifacts.py into smaller files
